Initial version allowing for instantiation of Operator
Fix various issues and add enhancements that allows for the service-assurance-operator to
start up from the OperatorHub. Focus was allowing for the execution of
operator-courier verify to allow for testing to execute.

Additionally, testing via operator-sdk scorecard was added to allow the CSV and
other components to execute. Not all issues identified by the test are
resulting in a pass, but we're far enough along now that a test can be executed
fully. A new configuration file was added to the repo to allow for things to be
reproduced.

Basic testing has been added to Travis-CI to enable the operator-courier verify test
to run so that we don't regress.

Final change was allowing for the SAF Cluster provided API to be shown in the OpenShift
UI so that someone can create an instance of SAF from the UI. No actual functionality
has been added at this time, as we've just stubbed out the framework.

Final change was migrating from service-assurance.infra.watch to just the top level
API of infra.watch.