Update CSV for latest changes
Update the CSV based on a better understanding of where the original values
are pulled from. The deploy/ directory contains the main source of truth and
the olm-catalog is then generated from the data involved there. For example, the
alm-example is populated from the deploy/crds/*_cr.yaml files. The permissions are
pulled from the values written in the deploy/roles.yaml, etc.

The update_csv.sh also cleans up the positioning of items of information based
on the formatting the operator-sdk olm-catalog gen-csv expects. Don't fight
it.