Refinement of playbooks prior to alpha release
Clean up the playbooks and get them in a reasonable shape for an alpha release
that others could use. Intent here is to get things into the first setup that allows
for an override of the various manifests that are created by default. Definition
of a default manifest has been included. Expectation is that the defined manifests
will be adjusted going forward.