Remove unnecessary manifest parameters
Tested some more and the issue with the Interconnect instance creation resulting
in a loop of the Service Assurance Operator was due to the missing sslProfile that
is automatically created as part of the interRouterListener creation. When I add
the openstack sslProfile to the list, that results in the list being out of sync
with what the Operator wants the Interconnect object to look like.

By specifically calling out the inter-router sslProfile we eliminate this race
condition that results in indefinite object updating within the Service Assurance
Operator.