Update statusDescriptors so it is valid again
Unfortunately I might have found a bug or usecase that is not yet documented. Our status.conditions
output is a list, and so far I have not found many examples of the status output being a list.

There does exist an option in the watches.yaml for manageStatus: false that would allow us to
manage the status ourselves. That might ultimately be the solution here, however is a task
for another day.

This change is to give us some information about what happened in the run, and to allow for
the status to show up. Unforunately at this point it will result in the operator-sdk
scorecard from passing fully, but that's more a nice-to-have option at this point.

Closes #17