Migrate to non-versioned container registry
Migrate the container images to a registry that doesn't include the CSV version
in the name. To do this, need to remove the old application registry of the same
desired name to serviceassurance-operator in order to free up the use of
service-assurance-operator for the container registry.

(The naming convensions used match our naming for delivery of the Smart Gateway Operator.)

Script for pushing latest CSV into the application registry has been updated to
match the new name and validated to have worked. The deploy/operator.yaml file has
been adjusted and the operator-sdk olm-catalog gen-csv has been executed to get the
generated olm-catalog contents up to date.