Migrate to using interRouterListeners by default
As of OSP16 QDR connections will use inter-router connections from controllers
to access the QDR for SAF server side. This change adjusts the default configuration
for OSP16 setup which is the target for SAF release.

Backwards compatibility with OSP13 environments can be maintained by passing
a custom interconnect_manifest parameter to SAO to either expose port 5671
as an edgeListener, or add an additional port to support both OSP13 and OSP16
simultaneously.

Intent is to backport the QDR mesh work to OSP13 so that both environments
connect the same way. We've decided to change the defaults to match our ideal
situation, which should make documentation and future development simpler.