Provide method to override SG container image target
Provide a way in the Ansible role to override the container image path and tag, while
using a default value of 'latest'. When creating a new stable release, a specific container
image version should be provided for both the Operator container image and (optionally) the
Smart Gateway container image tag. The CustomResource should be updated to point at a specific
smartgatewayContainerImageTag value to result in a deployed Smart Gateway of a released
version rather than 'latest'.