Tweaks to result in an OSP16 deployment
Changes to the infrared-openstack.sh script to result in the successful deployment of OSP16.0
with Ceph and Tempest added for fun and profit. Parameterizes several options that were static
before to allow for a more flexible script while assuming primarily defaults will be utilized.

Re-adds the internal knowledge because it's been discussed that leaking this doesn't seem to
be a huge issue(?). Can pull it out in a subsequent commit as it'll be killed in a squash.