Adjust how infrared is used in test scripts
Adjust the test script so that every step is a separate bash function. This will allow us some
additional controls in the future should it be wanted, but for now it's just allowing a new
environment variable to be passed called TEMPEST_ONLY accepting true/false values.

Originally I wanted this setup so that I could re-run the tempest tests to provide some
sort of load on the cloud for additional STF testing and possibly in the future running
a specific set of tests that might result in expected alarms etc firing off (which could
test the end-to-end transit).

Additinally the metrics-qdr-connectors.yaml has been renamed to better reflect the upstream
documentation, so the file is now stf-connectors.yaml.