Approaching POC near completion
Adds the setup_base.yml which does the supporting OperatorGroup, CatalogSources,
and supporting Subscriptions for an STF deployment.

Adds the deploy_stf.yml which gets an STF environment up and running just prior
to the creation of a ServiceTelemetry object which would fire off the rest of
the supporting components.

Once creation of an STF instance is running, next steps are to implement the
checks to validate that the environment is running.