Complete POC
Complete initial implementation of an ansible role that replaces the existing
bash scripts and CI deployment.

This commit adds the validation script which is a stripped version of deploy_ci.sh
with just the final environment validation left. Also completes the deployment
of STF by creating a ServiceTelemetry object.

Starts the clean up effort by removing old scripts that are now consumed by this
Ansible role.