Change approach to creating BuildConfigs
Change the approach to BuildConfigs by using the oc new-build configuration
to create a new BuildConfig from the local Dockerfile and then triggering a
start-build with --from-dir to make sure we're building against local source.

Adds a filter to the validation because you will always have the first image build
fail, and there is no suppression to the Builds being created via oc new-build.

Also fixes an issue with the sed line on the STO CSV, and adjusts the hack that
returns the latest CSV version from a directory list.