Add updated CI scripts for multi-repo coordination (#101)
* Add scaffolding for CI deployment

* Add methods to create builds

Creates build and stores the image path into a handy reference variable for later

* Add branch cloning logic

* Run new CI instead of old

Also adds a clean up action for the image building to the remove_stf.sh

* Adjust ci.yml

* Add debug info

* Debug

* Adjust playbook name and calling it

* Debug...?

* Debug:

* Don't fight CI

* More output when builds fail

* Result in local builds

* Drop debugging lines

* Approaching POC near completion

Adds the setup_base.yml which does the supporting OperatorGroup, CatalogSources,
and supporting Subscriptions for an STF deployment.

Adds the deploy_stf.yml which gets an STF environment up and running just prior
to the creation of a ServiceTelemetry object which would fire off the rest of
the supporting components.

Once creation of an STF instance is running, next steps are to implement the
checks to validate that the environment is running.

* Complete POC

Complete initial implementation of an ansible role that replaces the existing
bash scripts and CI deployment.

This commit adds the validation script which is a stripped version of deploy_ci.sh
with just the final environment validation left. Also completes the deployment
of STF by creating a ServiceTelemetry object.

Starts the clean up effort by removing old scripts that are now consumed by this
Ansible role.

* Change approach to creating BuildConfigs

Change the approach to BuildConfigs by using the oc new-build configuration
to create a new BuildConfig from the local Dockerfile and then triggering a
start-build with --from-dir to make sure we're building against local source.

Adds a filter to the validation because you will always have the first image build
fail, and there is no suppression to the Builds being created via oc new-build.

Also fixes an issue with the sed line on the STO CSV, and adjusts the hack that
returns the latest CSV version from a directory list.