Enhance CI automation (#106)
* Add ServiceTelemetry overrides

Allows ServiceTelemetry overrides to be expressed via Ansible extra-vars. Adds the four
main overrides you would expect in a ServiceTelemetry object, with appropriate defaults
set.

Will also allow passing in the service_telemetry_manifest as a whole object like what we
do with the Service Telemetry Operator.

* Allow for per-repo branch overrides

Allow for per-repo branch overrides for the Smart Gateway Operator and Smart Gateway
repositories via sgo_branch and sg_branch (respectively).

* Add functionality around quickstart.sh

Add some functionality that was replaced when I dropped the quickstart.sh. Adds
some of this functionality back in and also adds some new stuff.

* Fix syntax error

* Add back quickstart.sh

Adds back a quickstart.sh that simulates the same result as the old
quickstart.sh

* Better CSV modification support

Also adds some tags to make skipping over builds for testing much
easier.

* Debugging ci.yml firing

* Make sure namespace is set before using it

* Fix syntax and documentation

* Test locally first kids

* Drop CI debug lines

* Clean up working repo clones

* Copy CSV into working directory

On subsequent runs the in-place modification of the CSV can cause issues either
in the development environment, or re-runs of the CI system. Copying the CSV out
of the in-place repo into a working location, and then modifying in-place results
in a cleaner setup.

By doing the copy of the CSV files, we can drop the need to force clone the supporting
repositories.

Also cleans up some shell commands that were commented out now that they are being
dealt with via the replace module. Removes the extra commands added to ci.yml.