Support for Smart Gateway compile changes (#112)
* Support for Smart Gateway compile changes

Adds changes to the CI system to support the changes for Smart Gateway moving to
CentOS 8 and go mod. Was having some issues with building the Smart Gateway image
so this adds some additional logging output on the console. Also does some clean up
by removing the first build which is expected to fail. (Failure is expected because
on creation of the BuildConfig it will fire a Build, but it fails due to how the
Build is configured on auto-trigger. Subsequent builds pass.)

* print more container logs

* Drop build log --follow

Drop the --follow flag from the 'oc start-build' in case it is causes some sort of issue
in our ci-runner.

Co-authored-by: pleimer <pfbleimer@gmail.com>