Migrate to bundle format
Migrate to bundle format for the current release of Service Telemetry Operator.

The 1.1.0 directory remains for now as we'll migrate the repository to that shortly. We
need to release a bundle of our current release so we can migrate to the new CatalogSource
without changing artifacts, and then do an Operator upgrade from there.