Draft implementation of multi-cloud support
Provide a draft implementation of the multi-cloud support. Updates the CRD to account for recent
changes with inability to merge lists of objects easily. In return this adds new functionality
that will remove old SmartGateways should they be removed from the current list.

Adds the operator_sdk/util collection so that the k8s_status module is provided. Adds initial
support for a custom status notifying the admin that ephemeral storage is in place and is not
recommended. Will update statusDescriptors to make this obvious in the UI.

Removes unneeded component files and merges all logic into a single component_clouds.yml file.