Change naming format of SmartGateway objects
Changed the naming format of the SmartGateway objects because it was causing
the labels for the ConfigMap hash to overflow past the 63 char limit. In the
future we're probably going to have to add some extra checks in STO to make
sure it doesn't cause a naming length failure.

I've adjusted some of the naming to be less verbose overall as well.