Dictionary-based API and Multi-Cloud Support (#119)
* servicetelemetry.infra.watch update to v2alpha1

Defaults need to nest under servicetelemetry_defaults

Update CRD to new object-based schema

Adjust defaults and vars to match new CRD

[untested] Check in point of CRD conversion

Hit a milestone that I wanted to get checked into the repository. This includes some
additional changes in the CRD to better align the storage configuration story.

Starts to update a couple of templates for use with the new schema. Nothing is
tested yet since I expect the deployments to fail.

Update CI automation (#132)

Update the CI automation to account for new bundle format within
the Smart Gateway Operator and Service Telemetry Operator repositories.

[milestone] Backend deployments are working

Backend deployments are now working and debugging output of the list of clouds is now
working.

Empty storage object breaks Prometheus

* Adjust defaults for highAvailability

Adjust the setup of a nested dictionary defaults when highAvailable.enabled
is set to a value of true.

* Create block to manage HA variables

* Draft implementation of multi-cloud support

Provide a draft implementation of the multi-cloud support. Updates the CRD to account for recent
changes with inability to merge lists of objects easily. In return this adds new functionality
that will remove old SmartGateways should they be removed from the current list.

Adds the operator_sdk/util collection so that the k8s_status module is provided. Adds initial
support for a custom status notifying the admin that ephemeral storage is in place and is not
recommended. Will update statusDescriptors to make this obvious in the UI.

Removes unneeded component files and merges all logic into a single component_clouds.yml file.

* Update bundle CRD

* Bump CRD version to v1 and cleanup

Perform a bit of cleanup and bump the apiVersion of the CRD API to v1

* Only create SGs when supporting backend available

* Revert changes to CRD apiVersion v1

Revert the changes to the CRD apiVersion bump since it seems to be
incompatible with the current bundle setup with operator-sdk 0.17.2.

Catch up on current specDescriptor changes as well.

* Fix issue introduced in revert

* Non-functioning dual-API version support

* Set v1alpha1 descriptors to previous setup

* Add v1alpha1 backwards compatibility

* Update to operator-sdk 0.19, pass bundle validation

* Fix service monitor variable configuration

* Fix ansible linting and invalid variable usage

* Change naming format of SmartGateway objects

Changed the naming format of the SmartGateway objects because it was causing
the labels for the ConfigMap hash to overflow past the 63 char limit. In the
future we're probably going to have to add some extra checks in STO to make
sure it doesn't cause a naming length failure.

I've adjusted some of the naming to be less verbose overall as well.

* Don't double define SmartGateway object name

* Compact logic for SG naming

* Drop outdated test-local directory

* Environment changes for CI and smoketests

* ServiceTelemetry object renaming effort

* Adjust author email in ansible metadata

* Migrate 1.1.0 directory to manifests/

* Update ServiceTelemetry UI hints

Update the specDescriptors and statusDescriptors for the supported parameters in the new
ServiceTelemetry API.:

* Revert change in default example functionality

* Remove superfluous merge type setting in k8s module