[master merged] Add snmp traps container and config (#137)
* Add snmp traps container and config

as addition to the sto. It runs as a separate service on OCP
and will send out notifications via snmp to a given address.

* Move SNMP target to CRD

actually pull the deployment if snmp traps are selected

Unrelated: fix the operator group name to match the name in tests.

* Dedicated names for pods

* Finally fix the manifest issue

* Address review comments - go back to failing

* Break out the service into an additional manifest

- keep the snmp_traps manifest so that it can be overridden.
- Update the image location.

* Restore previous 1.0.2 csv

* Incorporate review feedback, fix typo.

* New CSV 1.1.0

* add snmp webhook to alertmanager config

Also add a smoketest

* Migrate to new CRD interface

* Update CRD and CR then generate bundle

* Use env var from operator.yaml for images

Use an environment variable to define the supporting images such as prometheus-webhook-snmp
so that we don't embed image paths directly in to the Ansible role. This is to support releases
where paths need to change or tags need to be pinned.

* Clean up to align to master branch

* Pass ansible linting

Co-authored-by: Matthias Runge <mrunge@redhat.com>