Add osp 16.1 deploy with sending metrics (#142)
* Add osp 16.1 deploy with sending metrics

to gnocchi and to stf.

Ceilometer interval is reduced to 5 secs. Also this configures
ceph daemons to send ceph metrics.

* Allow for environment file selection

Provides environment file selection between the stf and stf+gnocchi
configuration. Removes the PREFIX option which was causing deployment
issues. Drops the size of the controllers to a more reasonable size.

* Move changes from dir 16.1 to 16

* Update tests/infrared/16/stf-connectors.yaml.template

Co-authored-by: Leif Madsen <lmadsen@redhat.com>

* Update tests/infrared/16/stf-connectors.yaml.template

Co-authored-by: Leif Madsen <lmadsen@redhat.com>

* Update tests/infrared/16/stf-connectors.yaml.template

Co-authored-by: Leif Madsen <leif@leifmadsen.com>
Co-authored-by: Leif Madsen <lmadsen@redhat.com>