Bump ElasticSearch version to 7.10.2
Bump the ElasticSearch version in the template to 7.10.2, the last free version distributed
under the Apache v2 license (later versions are SSPL). The latest ECK version of 1.4 no longer
supports installing ElasticSearch 7.5.1 and results in a failed ElasticSearch object due to
7.5.1 being unsupported.

In the future this version value should be exposed in the ServiceTelemetry object so administrators
can set the desired version for their environment.