Update infrared-openstack.sh 13 script (#155)
* Add domain naming and STF enablement selector

Add OVERCLOUD_DOMAIN to uniquely identify this cloud in STF.

Add STF enablement selector to allow skipping STF enablement.

* Match controller cpu/mem with 16

* Align STF template to STF 1.0 multi-cloud documentation

Align the STF documentation we're using to provide naming functionality of the cloud when
deploying. The configuration of this file matches the multi-cloud configuration from the
Service Telemetry Framework 1.0 documentation.

* Update mirror configs

* Update tests/infrared/13/infrared-openstack.sh

* Align to latest OSP16 installation

Better align the installation script and the stf-connectors template to what
we're using for OSP16. Also adds in the ceilometer metrics configuration for
OSP13 which is now available.

* Fix linting

* Allow deployment to fully finish

Pass overcloud-templates option along with ceilometer-write-qdr-edge-only,
collectd-write-qdr-edge-only and custom outputs/stf-connectors.yaml which results
in a fully completed cloud. Without these options the Docker images file would
not be populated, and would error out when trying to deploy stf-connectors.yaml.