Handle Gnocchi enablement in infrared-openstack.sh
Allow enablement of gnocchi and aodh to act more like the STF enablement in the
infrared-openstack.sh script. Add ENABLE_GNOCCHI_CONNECTORS parameter, and set
it to true as we should probably be always testing Gnocchi and STF together.

Remove the old stf-connectors-gnocchi template and replace it with a condensed
version in gnocchi-connectors.yaml.

The file is currently static, but the script is set up like STF to provide a
method to add parameters in the future should anything be required.