Add support for smartgateway.infra.watch/v2 (#198)
* Add support for smartgateway.infra.watch/v2

Migrate to using the new smartgateway.infra.watch/v2 CRD in the Smart
Gateway Operator.

Related: https://github.com/infrawatch/smart-gateway-operator/pull/84
Related: rhbz#1959594

* Clear out old CRDs when running CI

When deleting the executing project the CRDs are not cleared out of system. This can
cause a conflict with the oc apply merge if the CRDs change significantly, like when
changing the apiVersion value.

Solution is to clear out the CRDs prior to doing a deployment with run-ci.yml.

* Fix config syntax for event application in sg-core

* Update smoketests

* Enable snmptrap support by default in run-ci

Enable snmptraps support by default in stf-run-ci role since the smoketest
expects it to be supported. This better aligns the stf-run-ci role and the
smoketest so that no extra gymnastics are required.

Deployment now looks like:

ansible-playbook ./build/run-ci.yaml
./tests/smoketest/smoktest.sh

* Drop Interconnect listener on 5672

Drop port 5672 listener on Interconnect since we're not longer using it. With the migration
to all sg-core + sg-bridge deployments means we no longer need this separate port.

* Ceilometer SG can be smartgateway_deployment_size

Set up the sg-core to use the smartgateway_deployment_size value
regardless of SG type since there is no longer use of legacy smart
gateway for telemetry.

* Add suggested namespace

* Fix timestamp config for sg-core