Adjust bundle Dockerfile to build for a bundle (#228)
Previously a change was mistakenly made to the wrong LABEL line which set the bundle
LABEL to false when the backport LABEL should have been set to false. Fixing the
Dockerfile.in for the bundle creation to disable the backport, but setting the build
to know this is a bundle image.

Invalid change was merged in 08a2ab5fb22a517ec7d07fac15206ee3cff92236
and this fixes that error.

Cherry picked from commit d22d156635adf2dc95db14b1f39a77f7d82863b8
Signed-off-by: Leif Madsen <lmadsen@redhat.com>