Drop custom spec and status descriptors
Drop the custom spec and status descriptors from the Service Telemetry
Operator CSV file. In the latest OpenShift versions, the UI hints are
automatically created based on the CRD. Removing the custom descriptors
will result in better synchronization between what is available in the
CRD and what is shown in the ServiceTelemetry configuration within the
OpenShift UI.

No changes are necessary to the documentation or example manifests to
support this change, and should be relatively transparent to the
adminstrator.

Resolves: STF-592