Add more defaults to populate the UI
Add more defaults to the CR example and generate via operator-sdk
generate bundle to update the CSV template. These changes will result in
the population of the default clouds environment that would be created
per official documentation. Results in items being populated in the UI
due to the dropped x-descriptors no longer being present to set the
defaults. Makes it more clear how the 'clouds' parameter should be set.