Drop custom spec and status descriptors (#271)
* Drop custom spec and status descriptors

Drop the custom spec and status descriptors from the Service Telemetry
Operator CSV file. In the latest OpenShift versions, the UI hints are
automatically created based on the CRD. Removing the custom descriptors
will result in better synchronization between what is available in the
CRD and what is shown in the ServiceTelemetry configuration within the
OpenShift UI.

No changes are necessary to the documentation or example manifests to
support this change, and should be relatively transparent to the
adminstrator.

Resolves: STF-592

* Add more defaults to populate the UI

Add more defaults to the CR example and generate via operator-sdk
generate bundle to update the CSV template. These changes will result in
the population of the default clouds environment that would be created
per official documentation. Results in items being populated in the UI
due to the dropped x-descriptors no longer being present to set the
defaults. Makes it more clear how the 'clouds' parameter should be set.