Limit cloud name to 10 characters
Limit the cloud name to 10 characters to try and keep the total length
limit to 19 characters, combined with the ServiceTelemetry object name +
cloud name which is combined as part of the Smart Gateway instantiation.
A combined length of greater than 19 characters between the
ServiceTelemetry object name and the cloud name will result in a pod
name length of greater than 63 characters, resulting in it not being
scheduled.