Fix issues with OCP_PROJECT in smoketest.sh
Fix issues with OCP_PROJECT not being set and verified properly in the
smoketest.sh script. Instead of relying on -z and word expansion,
explicitly set the OCP_PROJECT variable, substituting the value if
OCP_PROJECT has been previously set through and environment variable,
otherwise set the value to empty.

Adjust the 'if' statement to look for non-zero length to either change
to the project (implied by setting the project name), or use the
existing project returned from 'oc project -q' as the working project.

Additional fixes for the smoketest.sh script to bring it in compliance
with Shellcheck suggestions.