Provide method to install STF without dependencies
Change the way STF is deployed by only requiring the Smart Gateway API
as a dependency to allow Service Telemetry Operator to be deployed.

Update the main roles to only import when the appropriate API group is
available to the Operator. Remove the prometheus, AMQ interconnect, and
AMQ certmanager APIs from the requires list in the STO CSV.

Additionally, move the defaults to focus on deployment of the metrics
Smart Gateways and AMQ Interconnect bus by default to make it easier to
install a minimal STF environment that can be scraped by external
entities.

- Fix an issue with the default CSV alm-example that resulted in it
  reading the storageSelector as an object rather than a string, which
  caused issues with creating a ServiceTelemetry object via the web UI.

- Fix an issue with the default flavor value requested in the Loki
  backend configuration that was not valid based on the enum listed in
  the ServiceTelemetry CRD.

- Moves all backends to a state of false to allow deployment of STF by
  default as a telemetry-focused endpoint that can be scraped by any
  Prometheus instance desired by the administrator. The ability to
  request backend interfaces for ElasticSearch, Grafana, Alertmanager
  and Prometheus has been maintained, but is no longer enabled by
  default. STO now has the ability to deploy metrics Smart Gateways when
  a Prometheus backend is not available.