Don't parse ampersand in bash script
Fixes a bug where passing parameters from the ceilometer smoketest
entrypoint was not working because ampersands were being used without
quoting. This resulted in the passed parameters getting ignored and the
defaults used.

In this patch you can see we were passing 'metric' in the entrypoint
script. This was being ignored, but the smoketest worked because the
defaults resulting in the metering.sample address being used. Since this
is what the Service Telemetry Operator uses by default, we never saw the
incorrect address being specified, which would have resulted in the
smoketests failing.

By wrapping the parameters in single quotes, we can now modify this to
be a different address if necessary. The 'metric' address name has also
been updated to 'metering' as expected by the Service Telemetry
Operator.

Thanks to Paul Leimer with helping to track this down!