Smoketest and CI fixes (#282) (#285)
* Fix issues with OCP_PROJECT in smoketest.sh

Fix issues with OCP_PROJECT not being set and verified properly in the
smoketest.sh script. Instead of relying on -z and word expansion,
explicitly set the OCP_PROJECT variable, substituting the value if
OCP_PROJECT has been previously set through and environment variable,
otherwise set the value to empty.

Adjust the 'if' statement to look for non-zero length to either change
to the project (implied by setting the project name), or use the
existing project returned from 'oc project -q' as the working project.

Additional fixes for the smoketest.sh script to bring it in compliance
with Shellcheck suggestions.

* Don't parse ampersand in bash script

Fixes a bug where passing parameters from the ceilometer smoketest
entrypoint was not working because ampersands were being used without
quoting. This resulted in the passed parameters getting ignored and the
defaults used.

In this patch you can see we were passing 'metric' in the entrypoint
script. This was being ignored, but the smoketest worked because the
defaults resulting in the metering.sample address being used. Since this
is what the Service Telemetry Operator uses by default, we never saw the
incorrect address being specified, which would have resulted in the
smoketests failing.

By wrapping the parameters in single quotes, we can now modify this to
be a different address if necessary. The 'metric' address name has also
been updated to 'metering' as expected by the Service Telemetry
Operator.

Thanks to Paul Leimer with helping to track this down!
Cherry picked from commit 80623c00b69901673613cc1fc1aa422643d849ee