Clean up storageClass selector issues
Clean up storageClass selector issues which can cause PVCs to fail when
the backing driver does not support a selector. Adds a new parameter to
the stf-run-ci so that a storageClass value can be passed to the
deployment. Removes the storage_selector defaults from the
servicetelemetry role so that selectors aren't always defined when not
specified.