observabilityStrategy is now a top level parameter
Configure the observabilityStrategy parameter as a top level object. To
do this, it needs to be defined in roles/.../defaults/main.yml at the
top of the file (like cloudsRemoveOnMissing).

For top level parameters, if they are not a mapping, then they are not
defined in roles/.../vars/main.yml and can be defined directly in the
defaults configuration. Tasks then were updated to use the top-level
parameter name vs in the servicetelemetry_vars map.