Add observabilityStrategy option of none (#294)
* Add observabilityStrategy option of none

Add observabilityStrategy of none which will result in no observability
components being deployed. Previously this option was configured as
backends.deploymentStrategy which was confusing. This change moves
backends.deploymentStrategy to a top level object, and renames it to
observabilityStrategy.

Closes: STF-458

* Implement observability strategy for CI

Implement a method to deploy with observabilityStrategy: none with the
CI tooling.

* Fix syntax in ansible output

* observabilityStrategy is now a top level parameter

Configure the observabilityStrategy parameter as a top level object. To
do this, it needs to be defined in roles/.../defaults/main.yml at the
top of the file (like cloudsRemoveOnMissing).

For top level parameters, if they are not a mapping, then they are not
defined in roles/.../vars/main.yml and can be defined directly in the
defaults configuration. Tasks then were updated to use the top-level
parameter name vs in the servicetelemetry_vars map.

* Add missing default configuration

* Adjustments to control logic

* Check for certmanager API for QDR

Check for certmanager API interface for QDR deployment since it requests
the creation of a certificate. Remove a superfluous check for
certmanager as it's already verified in the outer block condition
checks.

* Add VALIDATION_SCOPE var for validation script

Add a scope control for validate_deployment.sh that allows for an
observabilityStrategy value of 'none' to still pass validation

* Add additional CI controls for observabilityStrategy none