Expose minimal set of bridge controls for SGs
Expose a minimal set of bridge controls for sg-bridge in the clouds
collectors configuration. Exposes the ringBufferCount, ringBufferSize
and verbose options. Sets the default values to an increased value of
16384 which seemed to be safe in testing and deal with most of the
buffer overflows that were resulting in corrupted messages being passed
to sg-core. The values can now be tuned to match the environment easily.

Discussion also referenced originally in
https://github.com/infrawatch/smart-gateway-operator/pull/91. This
change helps satisfy for changes in sg-core that allows it to receive
larger message sizes per https://github.com/infrawatch/sg-core/pull/74.