Add note about  default() is necessary for clouds
Add a note that makes it clear if the defaults for the clouds parameter
are added to, then be sure to implement a default() filter in the
templates that reference the clouds.* parameters.

The reason is that the clouds paramater is not updatable partially like
the rest of the structure. If an admin only partially defines the
parameters in the clouds list item, then it's possible for referenced
sub-params to be empty, and thus cause failure when the parameter is
referenced. This would go for all list objects.

In the case of Service Telemetry Operator, nested lists are only used in
the clouds parameter.