Adjustments for role_binding in SG for CI (#327)
* Adjustments for role_binding in SG for CI

Update the stf-run-ci playbooks to account for the namespace:
placeholder that has been added to role_binding.yaml as part of the
oauth-proxy work. Before this change, the deployment of SGO would fail
in the CSV due to a missing ClusterRoleBinding attaching to the
service-account in the appropriate namespace. This change causes the
role_binding.yaml to be updated to the installation namespace so that
the service account is referenced in the role binding properly.

Also sneak in a small change that reverts the local in place change to
the role_binding.yaml for the local repository.

* Fix where change reversion happens

Also fix a missing path separator in the git checkout -- command.

Cherry picked from commit 59ab1ca09d225384edc7fabbdd26fe95fa59c74d