Update to operator-sdk 1.11 (#350)
I have no idea if this is okay or not, but I observed that this is different from the master branch and prevents me from using `operator-sdk run bundle` from CI automation when working from this branch. I'm opening this PR just to test the change and discuss whether it's safe or not.