Disable generating certificates with Interconnect
The Interconnect Operator expects the certmanager.k8s.io API (CRD)
interface to exist in order to generate certificates. As that interface
is not available on OCP 4.10, we need to disable the generation of
certificates, and instead manage the creation of Issuers and
Certificates with Service Telemetry Operator directly via Certificate
Manager for Openshift, which uses a different API interface.