Initial implementation of certificates for Interconnect (#371)
* Initial implementation of certificates for Interconnect

Provide the initial configuration for the certificate creation using
Cert Manager for OpenShift. Also provides the ConfigMap for the SASL
database which is no longer being generated in the deployment.

* Make ansible-linter happy

* Fix invalid apiVersion reference for cert-manager

Fix the invalid reference to certmanager.k8s.io for Elasticsearch and
Interconnect certificate generation to use the appropriate apiVersion:
cert-manager.io API interface

* Add permission for access to cert-manager.io API

Add permissions to the Service Telemetry Operator role to have access to
the certificates and issuers resources within the cert-manager.io
interface. Update the generated CSV manifest using operator-sdk 0.19.4
using:

  operator-sdk generate bundle \
    --channels unstable \
    --default-channel unstable \
    --manifests \
    --metadata

* Minor README.md updates

* Add CRC start up commands to README

* Disable generating certificates with Interconnect

The Interconnect Operator expects the certmanager.k8s.io API (CRD)
interface to exist in order to generate certificates. As that interface
is not available on OCP 4.10, we need to disable the generation of
certificates, and instead manage the creation of Issuers and
Certificates with Service Telemetry Operator directly via Certificate
Manager for Openshift, which uses a different API interface.

* Add links to comments providing overview of certificate block