Remove OperatorHub.io as a CatalogSource (#394)
Remove the OperatorHub.io CatalogSource and instead use the
community-operators CatalogSource which is available with an OCP
installation. Ideally this will avoid some of the conflicts we've been
seeing in our CI environment. This is a short term fix as future
development will likely make use of Observability Operator to provide
the metrics data store and alert delivery mechanism.

(cherry-picked from commit e50f0573295847d238df547d07dab8a6e460892d)

Backporting this change in order to help with environment cleanup in
testing to avoid some timeouts we're seeing related to this
CatalogSource.